fix(#4365): hand off relay parser state at turn boundary
