+++ ARCHITECTURE.md
@@ -731,7 +731,8 @@ src/
 │   │   │   ├── idle_jsonl.rs
 │   │   │   ├── orphan_reclaim.rs
 │   │   │   ├── relay_format.rs
-│   │   │   └── task_notification_context.rs
+│   │   │   ├── task_notification_context.rs
+│   │   │   └── turn_parser.rs
 │   │   ├── session_runtime/
 │   │   │   ├── channel_routing.rs
 │   │   │   ├── restore_cwd.rs
